checkin_api/requirements.txt
Django==4.2.6
mysqlclient==2.2.0
djangorestframework==3.14.0
Pillow==10.0.1